% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fips}
\alias{fips}
\title{List of state fips codes - 50 states plus DC}
\format{A list of state fips codes}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}
}
\usage{
fips
}
\description{
Some small geographies in some Census APIs can only be used under a state hierarchy.
This is a list of fips codes that may be looped over to retrieve data for all states.
}
\examples{
fips
# [1]  1  2  4  5  6  8  9 10 11 12 13 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 44 45 46 47 48 49 50 51 53 54 55 56
}
\keyword{datasets}
