% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcensus_functions.R
\name{getFunction}
\alias{getFunction}
\title{Internal function: Get the API response, return a data frame}
\usage{
getFunction(apiurl, key, get, region)
}
\arguments{
\item{apiurl, }{key, get, region}
}
\description{
Internal function: Get the API response, return a data frame
}
\examples{
none
getFunction()
}
\keyword{internal}

