% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fips}
\alias{fips}
\title{List of state fips codes - 50 states plus DC}
\format{
A list of fips codes for 50 states and the District of Columbia
}
\usage{
fips
}
\description{
Some small geographies in some Census APIs can only be used under a state hierarchy.
This is a list of fips codes that may be looped over to retrieve data for all states.
}
\examples{
fips
}
\keyword{datasets}
