% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fips}
\alias{fips}
\title{List of state fips codes - 50 states plus DC and Puerto Rico}
\format{A list of state fips codes}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}
}
\usage{
fips
}
\description{
Since some geographies are only available under a state hierarchy, this can be helpful if you want to retrieve data for many states.
}
\keyword{datasets}
