% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcensus_functions.R
\name{getCensus}
\alias{getCensus}
\title{Retrieve Census data from a given API}
\usage{
getCensus(
  name,
  vintage = NULL,
  key = NULL,
  vars,
  region = NULL,
  regionin = NULL,
  time = NULL,
  year = NULL,
  date = NULL,
  period = NULL,
  monthly = NULL,
  show_call = FALSE,
  convert_variables = TRUE,
  category_code = NULL,
  data_type_code = NULL,
  naics = NULL,
  pscode = NULL,
  naics2012 = NULL,
  naics2007 = NULL,
  naics2002 = NULL,
  naics1997 = NULL,
  sic = NULL,
  ...
)
}
\arguments{
\item{name}{The programmatic name of your dataset, e.g. `timeseries/poverty/saipe`
or `acs/acs5`. See `listCensusApis()` for options.}

\item{vintage}{Vintage (year) of dataset, e.g. 2014. Not required for timeseries APIs.}

\item{key}{A Census API key, obtained at https://api.census.gov/data/key_signup.html.
If you have a `CENSUS_KEY` or `CENSUS_API_KEY` stored in your .Renviron file, getCensus()
will automatically use that key. Using a key is recommended but not required.}

\item{vars}{List of variables to get. Required.}

\item{region}{Geography to get.}

\item{regionin}{Optional hierarchical geography to limit region.}

\item{time, year, date, period, monthly}{Optional arguments used for some time series APIs.}

\item{show_call}{List the underlying API call that was sent to the Census Bureau.}

\item{convert_variables}{Convert likely numeric variables into numeric data.
Default is true. If false, results will be characters, which is the type returned by
the Census Bureau.}

\item{category_code, data_type_code, naics, pscode, naics2012, naics2007, naics2002, naics1997, sic}{Optional arguments used in economic data APIs.}

\item{...}{Other valid arguments to pass to the Census API. Note: the APIs are case sensitive.}
}
\description{
Retrieve Census data from a given API
}
\keyword{api}
