% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcensus_functions.R
\name{getCensus}
\alias{getCensus}
\title{Retrieve Census data from a given API}
\usage{
getCensus(apiurl, key, vars, region, time = NULL, date = NULL,
  period = NULL, monthly = NULL)
}
\arguments{
\item{apiurl}{Root URL for a Census API - see list at http://api.census.gov/data.html}

\item{key}{Your Census API key, gotten from http://api.census.gov/data/key_signup.html}

\item{vars}{List of variables to get}

\item{region}{Geograpy to get}

\item{time}{Optional argument used for some time series APIs}

\item{date}{Optional argument used for some time series APIs}

\item{period}{Optional argument used for some time series APIs}

\item{monthly}{Optional argument used for some time series APIs}
}
\description{
Heavily based on work by Nicholas Nagle, https://rpubs.com/nnnagle/19337
}
\examples{
acs_2014_api <- 'http://api.census.gov/data/2014/acs5'
myvars <- c("B01001_001E", "NAME", "B01002_001E", "B19013_001E", "B19001_001E", "B03002_012E")
df <- getCensus(acs_2014_api, key="YOURKEYHERE", vars=myvars, region="tract:*&in=state:06")

# Retrieve over 50 variables
myvars2 <- paste('B04004_', sprintf('\%03i', seq(1, 105)), 'E', sep='')
df <- getCensus(acs_2014_api, key="YOURKEYHERE", vars=myvars2, region="county*")

# Get time series data
saipe_api <- 'http://api.census.gov/data/timeseries/poverty/saipe'
saipe <- getCensus(saipe_api, key=censuskey, vars=c("NAME", "SAEPOVRT0_17_PT", "SAEPOVRTALL_PT"), region="state:*", time=2011)

# Loop over all states using fips list included in package
tracts <- NULL
for (f in fips) {
regionget <- paste("tract:*&in=state:", f, sep="")
temp <- getCensus(acs_2014_api, key=censuskey, vars=myvars, region=regionget)
tracts <- rbind(tracts, temp)
}
}
\keyword{api}

