% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcensus_functions.R
\name{getCensus}
\alias{getCensus}
\title{Retrieve data frame from the Census API}
\usage{
getCensus(apiurl, key, vars, region)
}
\arguments{
\item{apiurl}{Root URL for a Census API - see list at http://api.census.gov/data.html}

\item{key}{Your Census API key, gotten from http://api.census.gov/data/key_signup.html}

\item{vars}{List of variables to get}

\item{region}{Geograpy to get}
}
\description{
Heavily based on work by Nicholas Nagle, https://rpubs.com/nnnagle/19337
}
\examples{
myvars <- c("B01001_001E", "NAME", "B01002_001E", "B19013_001E", "B19001_001E", "B03002_012E")
acs_2014_api <- 'http://api.census.gov/data/2014/acs5'
myvars <- c("B01001_001E", "NAME", "B01002_001E", "B19013_001E", "B19001_001E", "B03002_012E")
df <- getCensus(acs_2014_api, key="YOURKEYHERE", vars=myvars, region="tract:*&in=state:06")
df <- getCensus(acs_2014_api, key="YOURKEYHERE", vars=myvars, region="county*")
}
\keyword{api}

