% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusapi_functions.R
\name{getCensus}
\alias{getCensus}
\title{Retrieve data frame from the Census API}
\usage{
getCensus(apiurl, key, vars, region)
}
\arguments{
\item{apiurl, }{key, vars, region}
}
\description{
Heavily based on work by Nicholas Nagle, https://rpubs.com/nnnagle/19337
}
\examples{
myvars <- c("B01001_001E", "NAME", "B01002_001E", "B19013_001E", "B19001_001E", "B03002_012E")
}
\keyword{api}

