% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcensus_functions.R
\name{getCensus}
\alias{getCensus}
\title{Retrieve Census data from a given API}
\usage{
getCensus(
  name,
  vintage = NULL,
  key = NULL,
  vars,
  region = NULL,
  regionin = NULL,
  time = NULL,
  year = NULL,
  date = NULL,
  period = NULL,
  monthly = NULL,
  show_call = FALSE,
  convert_variables = TRUE,
  category_code = NULL,
  data_type_code = NULL,
  naics = NULL,
  pscode = NULL,
  naics2012 = NULL,
  naics2007 = NULL,
  naics2002 = NULL,
  naics1997 = NULL,
  sic = NULL,
  ...
)
}
\arguments{
\item{name}{The programmatic name of your dataset, e.g. `timeseries/poverty/saipe`
or `acs/acs5`. See `listCensusApis()` for options.}

\item{vintage}{Vintage (year) of dataset, e.g. 2014. Not required for timeseries APIs.}

\item{key}{A Census API key, obtained at https://api.census.gov/data/key_signup.html.
If you have a `CENSUS_KEY` or `CENSUS_API_KEY` stored in your .Renviron file, getCensus()
will automatically use that key. Using a key is recommended but not required.}

\item{vars}{List of variables to get. Required.}

\item{region}{Geography to get.}

\item{regionin}{Optional hierarchical geography to limit region.}

\item{time}{Time period of data to get, used with time series APIs.}

\item{year, date, period, monthly, category_code, data_type_code, naics, pscode, naics2012, naics2007, naics2002, naics1997, sic}{Optional arguments used in timeseries data APIs.}

\item{show_call}{List the underlying API call that was sent to the Census Bureau.}

\item{convert_variables}{Convert likely numeric variables into numeric data.
Default is true. If false, results will be characters, which is the type returned by
the Census Bureau.}

\item{...}{Other valid arguments to pass to the Census API. Note: the APIs are case sensitive.}
}
\description{
Retrieve Census data from a given API
}
\examples{
\dontrun{
# Get total population and median household income for places (cities, towns, villages)
# in a single state from the 5-year ACS.
acs_simple <- getCensus(
  name = "acs/acs5",
  vintage = 2022,
  vars = c("NAME", "B01001_001E", "B19013_001E"),
  region = "place:*",
  regionin = "state:01")
head(acs_simple)

# Get all data from the B08301 variable group, "Means of Transportation to Work."
# This returns estimates as well as margins of error and annotation flags.
acs_group <- getCensus(
  name = "acs/acs5",
  vintage = 2022,
  vars = "group(B08301)",
  region = "state:*")
head(acs_group)

# Retreive 2020 Decennial Census block group data within a specific Census tract,
# using the regionin argument to precisely specify the Census tract, county,
# and state.
block_group <- getCensus(
	name = "dec/dhc",
	vintage = 2020,
	vars = c("NAME", "P1_001N"),
	region = "block group:*",
	regionin = "state:36+county:027+tract:220300")
head(block_group)

# Get poverty rates for children and for people of all ages over time using the
# Small Area Income and Poverty Estimates API
saipe <- getCensus(
  name = "timeseries/poverty/saipe",
  vars = c("NAME", "SAEPOVRT0_17_PT", "SAEPOVRTALL_PT"),
  region = "state:01",
  time = "from 2000 to 2022")
head(saipe)

# Get County Business Patterns data for a specific NAICS sector.
cbp_2016 <- getCensus(
 name = "cbp",
  vintage = "2016",
  vars = c("EMP", "ESTAB", "NAICS2012_TTL", "GEO_TTL"),
  region = "state:*",
  naics2012 = "23")
head(cbp_2016)
}
}
\keyword{api}
