% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcensus_functions.R
\name{getFunction}
\alias{getFunction}
\title{Internal function: Get the API response, return a data frame}
\usage{
getFunction(apiurl, key, get, region, regionin, time, date, period,
  monthly, category_code, data_type_code, naics, pscode, naics2012,
  naics2007, naics2002, naics1997, sic, ...)
}
\arguments{
\item{apiurl, }{key, get, region, time}
}
\description{
Internal function: Get the API response, return a data frame
}
\keyword{internal}
