% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{has_api_key}
\alias{has_api_key}
\title{Is there a saved API token?}
\usage{
has_api_key()
}
\description{
Is there a saved API token?
}
\examples{
has_api_key()

}
\seealso{
Other helpers: 
\code{\link{get_api_key}()}
}
\concept{helpers}
