% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{listCensusApis}
\alias{listCensusApis}
\title{Get dataset metadata on all available APIs as a data frame}
\usage{
listCensusApis()
}
\description{
Scrapes {http://api.census.gov/data.json} and returns a dataframe
that includes: title, name, vintage (where applicable), url, isTimeseries (binary),
temporal (helpful for some time series), description, modified date
}
\examples{
apis <- listCensusApis()
head(apis)
}
\keyword{metadata}
