% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{listCensusApis}
\alias{listCensusApis}
\title{Get useful dataset metadata on all available APIs as a data frame}
\usage{
listCensusApis()
}
\description{
Scrapes {https://api.census.gov/data.json} and returns a dataframe
that includes: title, description, name, vintage, url, dataset type, and other useful fields.
}
\examples{
apis <- listCensusApis()
head(apis)
}
\keyword{metadata}
