% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{listCensusApis}
\alias{listCensusApis}
\title{Get dataset metadata on all available APIs as a data frame}
\usage{
listCensusApis()
}
\description{
Scrapes {http://api.census.gov/data.html}
}
\examples{
listCensusApis()
}
\keyword{metadata}
