% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{listCensusApis}
\alias{listCensusApis}
\title{Return all available APIs as data frame}
\usage{
listCensusApis()
}
\description{
Return all available APIs as data frame
}
\examples{
listCensusApis()
}
\keyword{metadata}

