% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{listCensusMetadata}
\alias{listCensusMetadata}
\title{Get information about a specific API as a data frame}
\usage{
listCensusMetadata(
  name,
  vintage = NULL,
  type = "variables",
  group = NULL,
  variable_name = NULL
)
}
\arguments{
\item{name}{API programmatic name - e.g. acs/acs5. See list of names with listCensusApis().}

\item{vintage}{Vintage (year) of dataset. Not required for timeseries APIs}

\item{type}{Type of metadata to return, either "variables", "geographies",
"groups", or "values". Default is variables.}

\item{group}{An optional variable group code, used to return metadata for a specific group
of variables only. This field is not used in all APIs.}

\item{variable_name}{A name of a specific variable used to return metadata about that
variable. This field is not used in all APIs.}
}
\description{
Get information about a specific API as a data frame
}
\examples{
\dontrun{

# List the variables available in the Small Area Health Insurance Estimates.
sahie_variables <- listCensusMetadata(
  name = "timeseries/healthins/sahie",
  type = "variables")
 head(sahie_variables)

# List the geographies available in the 5-year 2020 American Community Survey.
acs_geographies <- listCensusMetadata(
  name = "acs/acs5",
  vintage = 2020,
  type = "geographies")
 head(acs_geographies)

# list the variable groups available in the 5-year 2020 American Community Survey.
acs_groups <- listCensusMetadata(
  name = "acs/acs5",
  vintage = 2020,
  type = "groups")
 head(acs_groups)

# List the value labels of the NAICS2017 variable in the 2020 County
# Business Patterns dataset.
cbp_naics_values <- listCensusMetadata(
  name = "cbp",
  vintage = 2020,
  type = "values",
  variable = "NAICS2017")
 head(cbp_naics_values)

# List of variables that are included in the B17020 group in the
# 5-year American Community Survey.
group_B17020 <- listCensusMetadata(
  name = "acs/acs5",
  vintage = 2017,
  type = "variables",
  group = "B17020")
 head(group_B17020)
}
}
\keyword{metadata}
