% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_functions.R
\name{listCensusMetadata}
\alias{listCensusMetadata}
\title{Get variable or geography metadata for a given API as a data frame}
\usage{
listCensusMetadata(apiurl, type)
}
\arguments{
\item{apiurl}{Root URL for a Census API}

\item{type:}{Type of metadata to return, either 'v' for variables or 'g' for geographies}
}
\description{
Get variable or geography metadata for a given API as a data frame
}
\examples{
vars2014 <- listCensusMetadata("http://api.census.gov/data/2014/acs5", "v")
}
\keyword{metadata}

