% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusapi_functions.R
\name{listCensusMetadata}
\alias{listCensusMetadata}
\title{Return list of variables or geographies available by api}
\usage{
listCensusMetadata(apiurl, type)
}
\arguments{
\item{apiurl, }{type}
}
\description{
Return list of variables or geographies available by api
}
\examples{
vars2014 <- listCensusMetadata("http://api.census.gov/data/2014/acs5", "v")
}
\keyword{metadata}

